fix the lossy nature of the timedelta -> interval conversion
